refactor(ssh): create module for ssh factory
